Extract schema to separate directory
