+++ lib/index.js
@@ -9,6 +9,7 @@ import Joi from "joi";
 
 import compile from "./compile";
 import { entries } from "./util/object";
+import constructorInputSchema from "./schemas/interlock-constructor-input.js";
 import loadAst from "./compile/modules/load-ast";
 import compileModules from "./compile/modules/compile";
 
@@ -56,45 +57,7 @@ function normalizeEntryPoints (entryPoints) {
  * @param   {String}    options.implicitBundleDest  The location to emit shared dependency bundles
  */
 export default function Interlock (options) {
-  const cwd = process.cwd();
-
-  const schema = Joi.object().keys({
-    srcRoot: Joi.string()
-      .required(),
-    destRoot: Joi.string()
-      .default(() => { return path.join(cwd, "dist"); }, "context + '/dist'"),
-    entry: Joi.object().pattern(/.*/, Joi.alternatives().try(
-      Joi.string(),
-      Joi.object().keys({
-        dest: Joi.string().required()
-      })
-    )),
-    split: Joi.object().pattern(/.*/, Joi.alternatives().try(
-      Joi.string(),
-      Joi.object().keys({
-        dest: Joi.string().required()
-      })
-    )),
-    includeComments: Joi.boolean()
-      .optional(),
-    sourceMaps: Joi.boolean()
-      .optional(),
-    plugins: Joi.array()
-      .optional(),
-    implicitBundleDest: Joi.string()
-      .default("[setHash].js")
-      .optional(),
-    extensions: Joi.array().items(Joi.string())
-      .default([".js", ".jsx", ".es6"])
-      .optional(),
-    ns: Joi.string()
-      .optional(),
-    context: Joi.string()
-      .default(cwd)
-      .optional()
-  }).or("entry", "split");
-
-  Joi.validate(options || {}, schema, (err, value) => {
+  Joi.validate(options || {}, constructorInputSchema, (err, value) => {
     if (err) { throw err; }
 
     // It is not possible to extend joi though it is an open discussion for a future version
